Create fizzbuzz.cpp
There are 10 or so examples of fizzbuzz included.
You will need to uncomment the one you want to use and make sure the rested are commented out.
The functions at the top you can leave uncommented. They won't effect the one's that don't use them.

I created this because I have seen to many people using examples of the first one.
There are always better options to using "else if".
It is a pretty clear indicator the person is still very new to programming or didn't think much on how to solve the problem or has a lot to learn when it comes to problem solving.
If statements aren't great for performance to start with the more you use the worse it gets. If you have a lot of if statements you probably should be using a switch or reconsidering how to look at the problem at hand.
Worse would be nesting if, if else and else if.
If you are doing that its a clear indicator you need to stop and think about the problem.

The examples: 
There is no best example on here. The best example is the one that fits your needs best.
However, there is a worst. That is the first example which many people use. Seriously don't do that.
There is a bit of a progression in how the examples change.
To an extent performance increases and the amount of code can decrease. However being able to maintain or expand those examples can be more difficult.
They are there to get you to see other ways at looking at how to solve problems like this.